pyproject.toml
[tool.poetry]
name = "navatech-multitenant-fastapi"
version = "0.1.0"
description = ""
authors = ["Subhomoy Roy Choudhury <subhomoyrchoudhury@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.12"
uvicorn = "^0.34.2"
sqlalchemy = "^2.0.41"
psycopg2-binary = "^2.9.10"
alembic = "^1.16.1"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.4.0"}
pyyaml = "^6.0.2"
pydantic-settings = "^2.9.1"
typeguard = "^4.4.2"
typer = "^0.16.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
